H180598Y Ruwarashe Garwe/flex-server-app/flex/flex/tasks/run_process.py
from celery.decorators import task

from major_damage.models import MajorDamageRun


@task(name="run_model")
def run_majour_damage_model(model_id):
    model = MajorDamageRun.objects.get(id=model_id)
    print(f"Run {model} now")
