H180598Y Ruwarashe Garwe/flex-server-app/requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
coverage==5.5
distlib==0.3.1
Django==3.1.11
django-easy-audit==1.3.0
django-simple-history==3.0.0
djangorestframework==3.12.4
filelock==3.0.12
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.0.2
MarkupSafe==2.0.1
packaging==20.9
pep8==1.7.1
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pybase64==1.1.4
pycodestyle==2.7.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-decouple==3.4
pytz==2021.1
redis==3.4.1
reportlab==3.5.67
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
vine==5.0.0
virtualenv==20.4.6
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==5.1.0
django-nested-admin==3.3.3